+++ public/src/components/tests/variants/testEditorVariantSummaryWebPreviewButton.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Button } from '@mui/material';
 import VisibilityIcon from '@mui/icons-material/Visibility';
-import { TestPlatform, TestType } from '../helpers/shared';
+import { TestPlatform, TestType } from '../../channelManagement/helpers/shared';
 import { getStage, Stage } from '../../../utils/stage';
 
 export type ArticleType = 'Standard' | 'Liveblog';
