+++ public/src/components/tests/variants/testNewVariantButton.tsx
@@ -4,7 +4,7 @@ import { makeStyles } from '@mui/styles';
 
 import AddIcon from '@mui/icons-material/Add';
 import useOpenable from '../../../hooks/useOpenable';
-import CreateVariantDialog from '../createVariantDialog';
+import CreateVariantDialog from '../../channelManagement/createVariantDialog';
 
 const useStyles = makeStyles(({ spacing, palette }: Theme) => ({
   button: {
