+++ public/src/components/tests/variants/testVariantEditorWithPreviewTab.tsx
