+++ public/src/components/tests/variants/variantDeleteButton.tsx
