+++ public/src/components/tests/variants/variantEditorCtaEditor.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Checkbox, FormControlLabel, Theme } from '@mui/material';
 import { makeStyles } from '@mui/styles';
-import { Cta } from '../helpers/shared';
+import { Cta } from '../../channelManagement/helpers/shared';
 import VariantEditorCtaFieldsEditor from './variantEditorCtaFieldsEditor';
 
 const useStyles = makeStyles(({ spacing }: Theme) => ({
